patchNaNs() and finiteMask() dispatched
