+++ modules/core/src/mathfuncs.cpp
@@ -1574,360 +1574,6 @@ bool checkRange(InputArray _src, bool quiet, Point* pt, double minVal, double ma
     return true;
 }
 
-#ifdef HAVE_OPENCL
-
-static bool ocl_patchNaNs( InputOutputArray _a, double value )
-{
-    int ftype = _a.depth();
-    int rowsPerWI = ocl::Device::getDefault().isIntel() ? 4 : 1;
-    ocl::Kernel k("KF", ocl::core::arithm_oclsrc,
-                     format("-D UNARY_OP -D OP_PATCH_NANS -D dstT=%s -D DEPTH_dst=%d -D rowsPerWI=%d",
-                            ftype == CV_64F ? "double" : "float", ftype, rowsPerWI));
-    if (k.empty())
-        return false;
-
-    UMat a = _a.getUMat();
-    int cn = a.channels();
-
-    // to pass float or double to args
-    if (ftype == CV_32F)
-    {
-        k.args(ocl::KernelArg::ReadOnlyNoSize(a), ocl::KernelArg::WriteOnly(a, cn), (float)value);
-    }
-    else // CV_64F
-    {
-        k.args(ocl::KernelArg::ReadOnlyNoSize(a), ocl::KernelArg::WriteOnly(a, cn), value);
-    }
-
-    size_t globalsize[2] = { (size_t)a.cols * cn, ((size_t)a.rows + rowsPerWI - 1) / rowsPerWI };
-    return k.run(2, globalsize, NULL, false);
-}
-
-#endif
-
-void patchNaNs( InputOutputArray _a, double _val )
-{
-    CV_INSTRUMENT_REGION();
-    CV_Assert( _a.depth() == CV_32F || _a.depth() == CV_64F);
-
-    CV_OCL_RUN(_a.isUMat() && _a.dims() <= 2,
-               ocl_patchNaNs(_a, _val))
-
-    Mat a = _a.getMat();
-    const Mat* arrays[] = {&a, 0};
-    int* ptrs[1] = {};
-    NAryMatIterator it(arrays, (uchar**)ptrs);
-    size_t len = it.size*a.channels();
-
-    if (_a.depth() == CV_32F)
-    {
-        Cv32suf val;
-        val.f = (float)_val;
-
-#if (CV_SIMD || CV_SIMD_SCALABLE)
-        v_int32 v_mask1 = vx_setall_s32(0x7fffffff), v_mask2 = vx_setall_s32(0x7f800000);
-        v_int32 v_val = vx_setall_s32(val.i);
-#endif
-
-        for (size_t i = 0; i < it.nplanes; i++, ++it)
-        {
-            int *tptr = ptrs[0];
-            size_t j = 0;
-
-#if (CV_SIMD || CV_SIMD_SCALABLE)
-            size_t cWidth = (size_t)VTraits<v_int32>::vlanes();
-            for (; j + cWidth <= len; j += cWidth)
-            {
-                v_int32 v_src = vx_load(tptr + j);
-                v_int32 v_cmp_mask = v_lt(v_mask2, v_and(v_src, v_mask1));
-                v_int32 v_dst = v_select(v_cmp_mask, v_val, v_src);
-                v_store(tptr + j, v_dst);
-            }
-            vx_cleanup();
-#endif
-
-            for (; j < len; j++)
-                if ((tptr[j] & 0x7fffffff) > 0x7f800000)
-                    tptr[j] = val.i;
-        }
-    }
-    else if (_a.depth() == CV_64F)
-    {
-        Cv64suf val;
-        val.f = _val;
-
-#if (CV_SIMD || CV_SIMD_SCALABLE)
-        v_int64 v_mnt_mask = vx_setall_s64(0x000FFFFFFFFFFFFF), v_exp_mask = vx_setall_s64(0x7FF0000000000000);
-        v_int64 v_val = vx_setall_s64(val.i);
-#endif
-
-        for (size_t i = 0; i < it.nplanes; i++, ++it)
-        {
-            int64* tptr = (int64*)ptrs[0];
-            size_t j = 0;
-
-#if (CV_SIMD || CV_SIMD_SCALABLE)
-            size_t cWidth = (size_t)VTraits<v_int64>::vlanes();
-            for (; j + cWidth <= len; j += cWidth)
-            {
-                v_int64 v_src = vx_load(tptr + j);
-                v_int64 vande = v_and(v_src, v_exp_mask);
-                v_int64 vandm = v_and(v_src, v_mnt_mask);
-                v_int64 ve, vm;
-
-                ve = v_eq(vande, v_exp_mask);
-                vm = v_ne(vandm, vx_setzero_s64());
-
-                v_int64 v_isnan = v_and(ve, vm);
-                v_int64 v_dst = v_or(v_and(v_isnan, v_val), v_and(v_not(v_isnan), v_src));
-                v_store(tptr + j, v_dst);
-            }
-            vx_cleanup();
-#endif
-
-            for (; j < len; j++)
-                if ((tptr[j] & 0x7FFFFFFFFFFFFFFF) > 0x7FF0000000000000)
-                    tptr[j] = val.i;
-        }
-    }
-}
-
-#if (CV_SIMD || CV_SIMD_SCALABLE)
-
-//TODO: make true SIMD code instead
-template <typename _Tp, int cn>
-int finiteMaskSIMD_(const _Tp *src, uchar *dst, size_t total)
-{
-    const int osize = 8;
-    int i = 0;
-    for (; i <= (int)total - osize; i += osize)
-    {
-        for (int j = 0; j < osize; j++)
-        {
-            bool finite = true;
-            for (int c = 0; c < cn; c++)
-            {
-                _Tp val = src[i * cn + j * cn + c];
-                finite = finite && !cvIsNaN(val) && !cvIsInf(val);
-            }
-            dst[i + j] = finite ? 255 : 0;
-        }
-    }
-
-    return i;
-}
-
-
-template <>
-int finiteMaskSIMD_<float, 1>(const float *src, uchar *dst, size_t total)
-{
-    const int osize = VTraits<v_uint8>::vlanes();
-    int i = 0;
-    for(; i <= (int)total - osize; i += osize )
-    {
-        v_uint32 vmaskPos = vx_setall_u32(0x7fffffff);
-        v_uint32 vmaskExp = vx_setall_u32(0x7f800000);
-        v_uint32 vv[4];
-        for (int j = 0; j < 4; j++)
-        {
-            v_uint32 vu = v_reinterpret_as_u32(vx_load(src + i + j*(osize/4)));
-            vv[j] = v_lt(v_and(vu, vmaskPos), vmaskExp);
-        }
-
-        v_store(dst + i, v_pack_b(vv[0], vv[1], vv[2], vv[3]));
-    }
-
-    return i;
-}
-
-
-template <>
-int finiteMaskSIMD_<float, 2>(const float *src, uchar *dst, size_t total)
-{
-    const int size8 = VTraits<v_uint8>::vlanes();
-    int i = 0;
-    for(; i <= (int)total - (size8 / 2); i += (size8 / 2) )
-    {
-        v_uint32 vmaskPos = vx_setall_u32(0x7fffffff);
-        v_uint32 vmaskExp = vx_setall_u32(0x7f800000);
-        v_uint32 vv[4];
-        for (int j = 0; j < 4; j++)
-        {
-            v_uint32 vu = v_reinterpret_as_u32(vx_load(src + i*2 + j*(size8 / 4)));
-            vv[j] = v_lt(v_and(vu, vmaskPos), vmaskExp);
-        }
-        v_uint8 velems = v_pack_b(vv[0], vv[1], vv[2], vv[3]);
-        v_uint16 vmaskBoth = vx_setall_u16(0xffff);
-        v_uint16 vfinite = v_eq(v_reinterpret_as_u16(velems), vmaskBoth);
-
-        // 2nd argument in vfinite is useless
-        v_store_low(dst + i, v_pack(vfinite, vfinite));
-    }
-
-    return i;
-}
-
-
-template <>
-int finiteMaskSIMD_<double, 1>(const double *src, uchar *dst, size_t total)
-{
-    const int size8 = VTraits<v_uint8>::vlanes();
-    int i = 0;
-    for(; i <= (int)total - (size8 / 2); i += (size8 / 2) )
-    {
-        v_uint64 vu[4];
-        for (int j = 0; j < 4; j++)
-            vu[j] = vx_load((const uint64_t*)src + i + j*(size8 / 8));
-
-        v_uint64 vmaskExp = vx_setall_u64(0x7ff0000000000000);
-        v_uint64 z = vx_setzero_u64();
-
-        v_uint64 vv[4];
-        for (int j = 0; j < 4; j++)
-        {
-            vv[j] = v_ne(v_and(vu[j], vmaskExp), vmaskExp);
-        }
-
-        v_uint8 v = v_pack_b(vv[0], vv[1], vv[2], vv[3], z, z, z, z);
-
-        v_store_low(dst + i, v);
-    }
-
-    return i;
-}
-
-template <>
-int finiteMaskSIMD_<double, 2>(const double *src, uchar *dst, size_t total)
-{
-    const int size8 = VTraits<v_uint8>::vlanes();
-    const int npixels = size8 / 16;
-    int i = 0;
-    for(; i <= (int)total - npixels; i += npixels )
-    {
-        v_uint64 vu = vx_load((const uint64_t*)src + i*2);
-
-        v_uint64 vmaskExp = vx_setall_u64(0x7ff0000000000000);
-        v_uint64 z = vx_setzero_u64();
-
-        v_uint64 vv = v_ne(v_and(vu, vmaskExp), vmaskExp);
-
-        v_uint8 velems = v_pack_b(vv, z, z, z, z, z, z, z);
-
-        v_uint16 vmaskBoth = vx_setall_u16(0xffff);
-        v_uint16 vfinite = v_eq(v_reinterpret_as_u16(velems), vmaskBoth);
-
-        // 2nd arg is useless
-        v_uint8 vres = v_pack(vfinite, vfinite);
-
-        for (int j = 0; j < npixels; j++)
-        {
-            dst[i] = v_get0(vres);
-            // 2nd arg is useless
-            vres = v_extract<1>(vres, vres);
-        }
-    }
-
-    return i;
-}
-
-#endif
-
-
-template <typename _Tp, int cn>
-static void finiteMask_(const _Tp *src, uchar *dst, size_t total)
-{
-    size_t i = 0;
-
-#if (CV_SIMD || CV_SIMD_SCALABLE)
-    i = finiteMaskSIMD_<_Tp, cn>(src, dst, total);
-#endif
-
-    for(; i < total; i++ )
-    {
-        bool finite = true;
-        for (int c = 0; c < cn; c++)
-        {
-            _Tp val = src[i * cn + c];
-            finite = finite && !cvIsNaN(val) && !cvIsInf(val);
-        }
-        dst[i] = finite ? 255 : 0;
-    }
-}
-
-#ifdef HAVE_OPENCL
-
-static bool ocl_finiteMask(const UMat img, UMat mask)
-{
-    int channels = img.channels();
-    int depth = img.depth();
-    int rowsPerWI = ocl::Device::getDefault().isIntel() ? 4 : 1;
-    ocl::Kernel k("finiteMask", ocl::core::finitemask_oclsrc,
-                  format("-D srcT=%s -D cn=%d -D rowsPerWI=%d",
-                         depth == CV_32F ? "float" : "double", channels, rowsPerWI));
-    if (k.empty())
-        return false;
-
-    k.args(ocl::KernelArg::ReadOnlyNoSize(img), ocl::KernelArg::WriteOnly(mask));
-
-    size_t globalsize[2] = { (size_t)img.cols, ((size_t)img.rows + rowsPerWI - 1) / rowsPerWI };
-    return k.run(2, globalsize, NULL, false);
-}
-
-#endif
-
-void finiteMask(InputArray _img, OutputArray _mask)
-{
-    CV_INSTRUMENT_REGION();
-
-    int channels = _img.channels();
-    int depth = _img.depth();
-    CV_Assert( channels > 0 && channels <= 4);
-    CV_Assert( depth == CV_32F || depth == CV_64F );
-    std::vector<int> vsz(_img.dims());
-    _img.sizend(vsz.data());
-    _mask.create(_img.dims(), vsz.data(), CV_8UC1);
-
-    CV_OCL_RUN(_img.isUMat() && _mask.isUMat() && _img.dims() <= 2,
-               ocl_finiteMask(_img.getUMat(), _mask.getUMat()));
-
-    Mat img = _img.getMat();
-    Mat mask = _mask.getMat();
-
-    const Mat *arrays[]={&img, &mask, 0};
-    Mat planes[2];
-    NAryMatIterator it(arrays, planes);
-    size_t total = planes[0].total();
-    size_t i, nplanes = it.nplanes;
-
-    for( i = 0; i < nplanes; i++, ++it )
-    {
-        const uchar* sptr = planes[0].ptr();
-        uchar* dptr = planes[1].ptr();
-
-        switch( depth )
-        {
-        case CV_32F:
-            switch (channels)
-            {
-            case 1: finiteMask_<float, 1>((const float*)sptr, dptr, total); break;
-            case 2: finiteMask_<float, 2>((const float*)sptr, dptr, total); break;
-            case 3: finiteMask_<float, 3>((const float*)sptr, dptr, total); break;
-            case 4: finiteMask_<float, 4>((const float*)sptr, dptr, total); break;
-            }
-            break;
-        case CV_64F:
-            switch (channels)
-            {
-            case 1: finiteMask_<double, 1>((const double*)sptr, dptr, total); break;
-            case 2: finiteMask_<double, 2>((const double*)sptr, dptr, total); break;
-            case 3: finiteMask_<double, 3>((const double*)sptr, dptr, total); break;
-            case 4: finiteMask_<double, 4>((const double*)sptr, dptr, total); break;
-            }
-            break;
-        }
-    }
-}
-
 } // namespace cv
